interviews/trivago/requirements.txt
voluptuous==0.11.5
pyyaml==3.13
